Rename userList component
