--- resources/js/components/usersList.js
@@ -1,17 +1,0 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-
-class UserData extends React.Component {
-
-}
-
-class UserList extends React.Component {
-
-}
-
-if (document.getElementById('userList')) {
-    RactDOM.render(
-        <UserList />,
-        document.getElementById('userList')
-    );
-}
